+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changelog
 
 ## 3.0.0
-- [feature] Allow form `onChange` and `handleFormData` data to change the forms control values in-place.
+- [feature] Allow form `onChange` and `handleFormData` functions to change the forms control values in-place.
 
 ## 2.17.0